Use a headless X server to run simulations.
deeprotor train-local now starts an X server before running the
simulation and tears it down afterwards. With this, it's possible to run
everything off the command line.

To access the GUI, connect to the VM using a VNC viewer pointed at
localhost:5900. The included Mac VNC viewer doesn't seem to work, but
VNC Connect by RealVNC works well.

If running on a remote machine, you can use ssh to port forward the VNC
connection and access the GUI remotely.